Fix math typo on website but the article still needs another read for sure. I was superrrrr exhausted and probably smoking weed when I wrote that...

+++ packages/rust/conundrum/src/parsers/conundrum/logic/number/conundrum_number.rs
@@ -1,8 +1,10 @@
 use serde::Serialize;
 
+#[typeshare::typeshare]
 #[derive(Debug, Serialize)]
+#[serde(tag = "tag", content = "content")]
 pub enum ConundrumNumber {
-    Int(i64),
+    Int(i128),
     Float(f64),
 }
